Modified rollcall macros to allow backend functions to be called with arguments passed by reference and by value.
Passing by reference is is now the automatic behaviour when a
variable is given as an argument to GET_BE_RESULT, as in:
GET_BE_RESULT(backendfuncname, myVariable).

Passing by value now only happens automatically if an explicit
value is given, as in:
GET_BE_RESULT(backendfuncname, 5).

To pass a variable by value, one now needs to explicitly
indicate that this is the desired behaiour, as in:
GET_BE_RESULT(backendfuncname, byVal(myVariable)).

	modified:   Backends/include/backend_libfirst.hpp
	new file:   Backends/lib/libfirst.cpp
	modified:   Backends/lib/libfirst.so
	modified:   Core/src/gambit_example.cpp
	modified:   ExampleBit_B/include/ExampleBit_B_rollcall.hpp
	modified:   ExampleBit_B/src/ExampleBit_B.cpp
	modified:   Utils/include/module_macros.hpp
	modified:   Utils/include/util_functions.hpp