I've added three new features to ScannerBit: 1)  The ability to share parameter across different models. 2)  The ability to fix a parameter to a specific value. 3)  The ability to redirect output relating to a specific model to a file or an x-terminal (buggy).
sharing parameters:

Let's say that you have two models that are declared in the ini-file as

model_0:
   param_0:
   param_1:

model_1:
   param_2:
   param_3:

The scanner sees 4 parameters: [model_0::param_0, model_0::param_1, model_1::param_2, model_1::param_3].  Note that the scanner sees the parameters keys in the form model::parameter.  Now what if param_0 and param_2 are the same parameter?  Well, these two can be joined by the key word "same_as:":

model_0:
   param_0:
   param_1:

model_1:
   param_2:
      same_as: model_0::param_0
   param_3:

Now the Scanner sees only three parameters:  [model_1::model_0::param_0, model_0::param_1, model_1::param_3].  But the model_1::model_0::param_0 parameter inputs into both the model_0::param_0 and model_1::param_2 parameters.  Also, not the the format for shared parameter names is model_shared::model::parameter.

You can also connect more than two parameter.  The following

model_0:
   param_0:

model_1:
   param_1:
      same_as: model_0::param_0

model_2:
   param_3:
      same_as: model_1::param_1

will be seen to the scanner as one parameter:  [model_2::model_1::model_0::param_0].  Also, if you do not include the "::", the scanner will assume that the parameters have the same name in both models.

Fixing values:

You can now fix the input value of a specific variable by using the "fixed_value" keyword.  For example, the entry

model:
   parameter:
      fixed_value: 0

Will fix the value of parameter to 0.  And as a consequence, the parameter "model::parameter" will not be seen by the scanner.

redirecting output:

You can now redirect output for a specific model to a file or x terminal with the ini-file entry:

enable_redirect: true

redirect_output:
   model: file

if "file" is "xterm", then the output will be directed to a x terminal.  Right now, this only works for one model, but it will soon work for more.

I've also moved all of the output to the "printErrors()" function.  This function is print errors, if there are any, and return true if the errors are fatal (false if not).