Implmented module function level ini-options
The "options" entry in observable or auxiliaries section is read into an
instance of the IniParser::Options class (previously only used for the prior
section), and a copy of this class is stored in the myOptions member variable
of the functor.

Note: myOptions is not yet available in the Pipes namespace, where it probably
belongs.

	modified:   Core/include/graphs.hpp
	modified:   Core/include/yaml_parser.hpp
	modified:   Core/src/graphs.cpp
	modified:   Utils/include/functors.hpp
	modified:   Utils/src/functors.cpp