base_functions now use boost::shared_ptr's.
Also, apparently, I did not have to force the 72 column limit.
The -ffixed-line-length-none flag was not being used since the
FOPT variable in the nulike makefile was not being used (it was
overshaddowed by the definition in the main makefile).  Renaming
it to FOPT2 seemed to fix it.  So, now I'll have to revert to
the old files.  Oh well, live and learn.