Made backends, modules and models excludable from the cmake command line.  Just do
 cmake -Ditch="blah1;blah2;etc" ..

to exclude all model/module/backends with names beginning with the
strings "blah1", "blah2" and "etc. For example,

 cmake -Dith="Coll;Dark;Example;nulike" ..

would exclude
modules ColliderBit
        DarkBit
        ExampleBit_A (including its standalone example)
        ExampleBit_B
as well as all versions of backend nulike.  Version 1.0.x would be
specified by saying nulike_1_0_ instead.

Also added a check for ROOT, and disabled Delphes and ColliderBit
automatically if it is not found.

	modified:   ../CMakeLists.txt
	new file:   ../cmake/FindROOT.cmake
	modified:   ../cmake/contrib.cmake
	new file:   ../cmake/executables.cmake
	modified:   ../cmake/gambit.cmake
	new file:   ../cmake/optional.cmake