Bug fix dependency resolver
There were some strange problems (pointed out by JC) when passing options to
module functions (both backend initialization functions and normal ones).  Took
some time to figure out what is going on.

The story is the following: getAuxiliaries() returns a constant reference to
some ObservablesType that contains the parsed information from the yaml
"Auxiliaries" section.  I used this to get a copy of this ObservablesType in
the following way:

    ObservablesType x = getAuxiliaries();

A few lines later, I use the same variable to get the entries from the ObsLike
section with:

    x = getObservables();

This copies the correct information to x, but at the same time overwrites a
small part of the initial auxiliaries entry inside the object with member
function getAuxiliaries() (namely the options for some of the auxiliaries
entries).  I don't know how that can happen.  Part of ObservablesType is
YAML::Node, which might contain references to stuff that is not copied when
doing the above copy construction of x but just referenced to.  Anyway, I
replaced now everything with

    const ObservablesType & x = getAuxiliaries();
    const ObservablesType & y = getObservables();

etc, which removed the problem.