"Beta" version of Standard Model parameter access added
Usage instructions:

You will ultimately need an object of type Spectrum*, and will get information
out of it in the same way as you do for say the MSSM. There are currently two
ways to get this object: methods (1) and (2) apply if you just want the
Standard Model stuff, while method (3) applies mostly if you need to access
MSSM information as well.

(1) In the relevant rollcall header, add a dependency on the capability
SM_spectrum, of type "const Spectrum*" (note: all Spectrum* need the const
keyword):
   >> DEPENDENCY(SM_spectrum, const Spectrum*)

You can then retrieve the dependency:
   >> const Spectrum* SM = *Dep::SM_spectrum;

get pole masses:
   >> SM->phys.get_Pole_Mass("Z")

(possible keys: W,Z,tau,t,b (plus aliases). The definition of b is a little
sketchy I think)

get MSbar masses
   >> SM->runningpars.get_mass_parameter("u")
(possible keys: d,u,s,c,b,t,e,mu,tau (plus aliases))

EM and strong couplings:
  >> SM->runningpars.get_dimensionless_parameter("alpha")
  >> SM->runningpars.get_dimensionless_parameter("alphaS")

set the scale of the above to X (must be below mTop) using
  >> SM->runningpars.RunToScale(X);

and that's it so far. Note, though, that because the type is const, you cannot
do the running on the original object. You have to make a copy first, like so:

  >> std::unique_ptr<Spectrum> spec2 = spec->get_UV()->clone();

(2) SLHA2 format SMINPUTS parameters. Add a dependency on SMINPUTS of type
SMInputs:

DEPENDENCY(SMINPUTS, SMInputs)

This is just a simple struct containing everything in the SMINPUTS block of an
SLHA2 file. For the list of data members, see
"Utils/include/gambit/Utils/SMplusUV.hpp"
(it is just everything in the SMINPUTS block)

(3) Get both of the above objects, plus the MSSM spectrum, all at once. Add a
dependency on MSSM_spectrum of type "const SMplusUV*". e.g.

  >> DEPENDENCY(MSSM_spectrum, const SMplusUV*)

Retrieve it:
const SMplusUV* matched_spectra = *Dep::MSSM_spectrum;

Pull out each of the above objects:

  >> SMInputs sminputs = matched_spectra->get_SMINPUTS();

  >> std::unique_ptr<Spectrum> SM = matched_spectra->clone_SM();
  >> const Spectrum* SM = matched_spectra->get_SM();

  >> std::unique_ptr<Spectrum> MSSM = matched_spectra->clone_UV();
  >> const Spectrum* MSSM = matched_spectra->get_UV();

(the cloned version are (new) copies and can be run; the "get" versions are
just pointers to the originals and cannot).

and then do all the stuff described above (and of course the usual to the MSSM
version as well).

FINAL NOTE: You will have to be scanning the model StandardModel_SLHA2 in order
for this stuff to be usable. See SpecBit.yaml for use, but basically just add

Parameters:
  StandardModel_SLHA2: !import StandardModel_SLHA2_defaults.yaml

to your yaml file, if you are happy with some defaults from a random SLHA
benchmark point for now. Will make a version corresponding to current PDG soon.

	modified:   DecayBit/include/gambit/DecayBit/DecayBit_rollcall.hpp
	modified:   DecayBit/src/DecayBit.cpp
	modified:   MSSM25_test.yaml
	modified:   Models/include/gambit/Models/models/StandardModel_SLHA2.hpp
	modified:   SpecBit/include/gambit/SpecBit/SpecBit_rollcall.hpp
	modified:   SpecBit/include/gambit/SpecBit/SpecBit_tests.hpp
	modified:   SpecBit/src/SpecBit.cpp
	modified:   SpecBit/src/SpecBit_tests.cpp