Fixed bugs in hdf5combine, offsets were not computed correctly when merging data into a previous combined output file.
This also reminded me that the "batching" modifications were not
yet made to the Auxilliary dataset combination. This means that
combinations where auxilliary datasets exist in lots of files
(Possibly TWalk? not sure) are vulnerable to a previous issue I
saw where the OS won't let you open too many files at once.

Actually I guess it must not be an issue for TWalk since it has
been working in my tests, but some scanners in the future might
trigger this problem.

	modified:   Printers/include/gambit/Printers/printers/hdf5printer/hdf5_combine_tools.hpp
	modified:   Printers/src/printers/hdf5printer/hdf5_combine_tools.cpp
	modified:   Utils/include/gambit/Utils/new_mpi_datatypes.hpp