Double whammy bugfix for 1) MPI shutdown messages and 2) HDF5Reader object
Regarding 1: To send and clean up shutdown messages I was previously doing
 Isend (to all)
 Wait (for all Isends to complete)
 Recv (all Isend messages)
But this is technically wrong and could theoretically lead to deadlocks, since
it is possible that Wait doesn't complete until the corresponding Recv occurs.
(in practice it seems to have been fine since the messages were small, and
probably they were sent in 'eager' mode). Anyway the other way around is
theoretically safe always, i.e.
 Isend (to all)
 Recv (all Isend messages)
 Wait (for all Isends to complete)
This is because the Recvs don't need the Waits to complete. So I have changed
this ordering.

Regarding 2: The HDF5Reader object did not check if the current pointID/MPIrank
pair (i.e. during iteration over the datasets) was valid, meaning that a chunk
of points with invalid MPIranks or pointIDs would crash it. This is now fixed.

	modified:   Printers/include/gambit/Printers/basebaseprinter.hpp
	modified:   Utils/src/signal_handling.cpp