Python link line correction
As discovered when installing on CSD3, if one has set PYTHON_LIBRARY for
installing pybind11, the final link line has -lpython2.7, which points
to the system version, rather than the version pointed to by
PYTHON_LIBRARY. If these versions are sufficiently different (e.g.
2.7.16 vs 2.7.5), then this causes a link error and gambit cannot
compile. This addition is intended as a conversation starter for a pull
request.